refactor(ui): extract stream-part rendering into chrome/thread/parts.tsx
Fifth extraction (Task 1): ThreadPart (text/caret/markdown, assistant
files, build beats, jailed generated-view app card), ThreadApprovals
(ENG-216 synthesis + ENG-205 morph + guard-first decide), and
ThreadConnectRequests move out of vendo-thread.tsx; UserText rides
along. Context values now read via useVendoContext at the new sites.
Behavior-preserving.
